# Delay checking prefix index until we need the result