--- a/benchmark/index.js
+++ b/benchmark/index.js
@@ -262,66 +262,77 @@ test('load two indexes concurrently', (t) => {
   const alice = ssbKeys.loadOrCreateSync(path.join(dir, 'secret'))
   const bob = ssbKeys.loadOrCreateSync(path.join(dir, 'secret-b'))
 
+  function waitForFile(filename, cb) {
+    fs.exists(filename, function(exists) {
+      if (exists) cb()
+      else setTimeout(() => waitForFile(filename, cb), 250)
+    })
+  }
+
   db.onReady(() => {
     const done = multicb({ pluck: 1 })
     const start = Date.now()
+    const contact_index_filename = path.join(indexesDir, 'type_contact.index')
 
-    db.indexes['type_contact'] = {
-      offset: 0,
-      bitset: new TypedFastBitSet(),
-      lazy: true,
-      filepath: path.join(indexesDir, 'type_contact.index'),
-    }
+    // we need to wait for the other tests to write the index file
+    waitForFile(contact_index_filename, () => {
+      db.indexes['type_contact'] = {
+        offset: 0,
+        bitset: new TypedFastBitSet(),
+        lazy: true,
+        filepath: contact_index_filename,
+      }
 
-    query(
-      fromDB(db),
-      where(
-        or(
-          equal(seekType, 'contact', { indexType: 'type' }),
-          equal(seekAuthor, alice.id, {
-            indexType: 'author',
-            prefix: 32,
-            prefixOffset: 1,
-          }),
-          equal(seekAuthor, bob.id, {
-            indexType: 'author',
-            prefix: 32,
-            prefixOffset: 1,
-          })
-        )
-      ),
-      toCallback(done())
-    )
-
-    query(
-      fromDB(db),
-      where(
-        or(
-          equal(seekType, 'contact', { indexType: 'type' }),
-          equal(seekAuthor, alice.id, {
-            indexType: 'author',
-            prefix: 32,
-            prefixOffset: 1,
-          }),
-          equal(seekAuthor, bob.id, {
-            indexType: 'author',
-            prefix: 32,
-            prefixOffset: 1,
-          })
-        )
-      ),
-      toCallback(done())
-    )
-
-    done((err) => {
-      if (err) t.fail(err)
-      const duration = Date.now() - start
-      t.pass(`duration: ${duration}ms`)
-      fs.appendFileSync(
-        reportPath,
-        `| Load two indexes concurrently | ${duration}ms |\n`
+      query(
+        fromDB(db),
+        where(
+          or(
+            equal(seekType, 'contact', { indexType: 'type' }),
+            equal(seekAuthor, alice.id, {
+              indexType: 'author',
+              prefix: 32,
+              prefixOffset: 1,
+            }),
+            equal(seekAuthor, bob.id, {
+              indexType: 'author',
+              prefix: 32,
+              prefixOffset: 1,
+            })
+          )
+        ),
+        toCallback(done())
       )
-      t.end()
+
+      query(
+        fromDB(db),
+        where(
+          or(
+            equal(seekType, 'contact', { indexType: 'type' }),
+            equal(seekAuthor, alice.id, {
+              indexType: 'author',
+              prefix: 32,
+              prefixOffset: 1,
+            }),
+            equal(seekAuthor, bob.id, {
+              indexType: 'author',
+              prefix: 32,
+              prefixOffset: 1,
+            })
+          )
+        ),
+        toCallback(done())
+      )
+
+      done((err) => {
+        if (err) t.fail(err)
+        const duration = Date.now() - start
+        t.pass(`duration: ${duration}ms`)
+        fs.appendFileSync(
+          reportPath,
+          `| Load two indexes concurrently | ${duration}ms |\n`
+        )
+        t.end()
+      })
     })
   })
 })