--- a/benchmark/index.js
+++ b/benchmark/index.js
@@ -14,11 +14,13 @@ const {
   query,
   fromDB,
   where,
+  and,
   or,
   equal,
   count,
   toCallback,
   toPullStream,
+  startFrom,
   paginate,
 } = require('../operators')
 const { seekType, seekAuthor, seekVoteLink } = require('../test/helpers')
@@ -499,6 +501,46 @@ test('query a prefix map (second run)', (t) => {
             )
           )
         })
+      })
+    )
+  })
+})
+
+test('paginate ten results', (t) => {
+  const alice = ssbKeys.loadOrCreateSync(path.join(dir, 'secret'))
+  const bob = ssbKeys.loadOrCreateSync(path.join(dir, 'secret-b'))
+
+  db.onReady(() => {
+    const start = Date.now()
+    pull(
+      query(
+        fromDB(db),
+        where(
+          and(
+            equal(seekType, 'contact', { indexType: 'type' }),
+            equal(seekAuthor, alice.id, {
+              indexType: 'author',
+              prefix: 32,
+              prefixOffset: 1,
+            })
+          )
+        ),
+        startFrom(0),
+        paginate(10),
+        toPullStream()
+      ),
+      pull.take(1),
+      pull.collect((err, msgs) => {
+        if (err) t.fail(err)
+        const duration = Date.now() - start
+        if (msgs[0].length !== 10)
+          t.fail('msgs.length is wrong: ' + msgs.length)
+        t.pass(`duration: ${duration}ms`)
+        fs.appendFileSync(
+          reportPath,
+          `| paginate 10 results | ${duration}ms |\n`
+        )
+        t.end()
       })
     )
   })