I was looking at the [performance of getting messages for a specific author](https://github.com/arj03/ssb-sync-test) and remembered there was one thing that I wanted to fix about prefix indexes, namely that it should be a bit more lazy about checking the messages it needs. Basically right now if you do query of: *give me the latest 10 messages from a specific author* it has to check all messages for that author to make sure that they match the prefix before it can sorted and results returned.

:warning: this changes the core structure of queries so should be reviewed carefully :warning: 

I made several rounds of this and the tests + benchmarks we have were quite useful in making sure that the code still works as before. I added an extra test to make sure that it works like I expected. Also this PR fixes a bug in the benchmarking code where we did not wait for an index to be written to disc before continuing and the bug was masked by the old code having to get a ton of messages from disc.

Benchmark results:

Before

```
db query warmup (10319): 2.985s
db query author all (8032): 959.953ms
db query author post msgs (2095): 617.328ms
db query author 10 msgs: 198.095ms
```

After:

```
db query warmup (10319): 1.013s
db query author all (8032): 760.335ms
db query author post msgs (2095): 236.625ms
db query author 10 msgs: 26.373ms
```