@staltz 2021-04-06
I'm still taking a look at this and wrapping my head around it, but one feedback is that we should probably use `Map` for the filters dict, it might improve performance a lot. To iterate over the keys, we should probably do `for (let key of map.keys())`

See these benchmarks I made: https://jsbench.me/2pkn5xz28v/1 (Left is Firefox, right is Chrome V8 and also what runs in Node.js)

![Screenshot from 2021-04-06 14-32-57](https://user-images.githubusercontent.com/90512/113704687-25c0c680-96e5-11eb-9028-a60322f938c3.png)