@arj03 2021-04-06
Ohh, the map was a good one:

MAP:

```
db query warmup: 576.999ms
db query author all: 491.68ms
db query author post msgs: 139.234ms
db query author 10 msgs: 13.678ms
```

PR:

```
db query warmup: 611.244ms
db query author all: 495.455ms
db query author post msgs: 153.325ms
db query author 10 msgs: 18.665ms
```