@Barbarrosa 2021-04-08
I was curious, so I ran the benchmark from #143 both before and after this change. Here are the results I got.  I read this as showing much leaner memory usage, but with some cost to speed.

I ran this after incorporating something similar to the benchmark bugfix in #142.

## Before

| Part | Speed | Heap Change | Samples |
|---|---|---|---|
| Load core indexes | 0.81ms ± 0.01ms | 435 B ± 33290 B | 418 |
| Query 1 big index (1st run) | 628.55ms ± 2.75ms | -241.92 kB ± 17.81 kB | 5 |
| Query 1 big index (2nd run) | 220.56ms ± 10.07ms | -284.24 kB ± 0 kB | 5 |
| Count 1 big index (3rd run) | 0.13ms ± 0ms | -467.72 kB ± 7.16 kB | 5 |
| Create an index twice concurrently | 535.85ms ± 36.27ms | 1.28 MB ± 0.04 MB | 5 |
| Query 3 indexes (1st run) | 648.78ms ± 3.79ms | 2.38 MB ± 0 MB | 5 |
| Query 3 indexes (2nd run) | 181.91ms ± 5.68ms | -65.67 kB ± 0 kB | 5 |
| Load two indexes concurrently | 336.02ms ± 15.4ms | 13.41 MB ± 0 MB | 5 |
| Paginate 20000 msgs with pageSize=5 | 8894.7ms ± 3624.68ms | -101.32 kB ± 0 kB | 5 |
| Paginate 20000 msgs with pageSize=500 | 405.45ms ± 32.54ms | 2.16 kB ± 0 kB | 5 |
| Query a prefix map (1st run) | 177.74ms ± 5.54ms | 5.03 MB ± 0.04 MB | 5 |
| Query a prefix map (2nd run) | 7.55ms ± 0.53ms | 2.49 MB ± 0.03 MB | 5 |

## After

| Part | Speed | Heap Change | Samples |
|---|---|---|---|
| Load core indexes | 0.81ms ± 0.01ms | 452 B ± 9951 B | 410 |
| Query 1 big index (1st run) | 623.05ms ± 5.11ms | -750.39 kB ± 45 kB | 5 |
| Query 1 big index (2nd run) | 228.82ms ± 10.64ms | -6.49 kB ± 42.09 kB | 5 |
| Count 1 big index (3rd run) | 0.13ms ± 0ms | -1.58 MB ± 0 MB | 5 |
| Create an index twice concurrently | 507.61ms ± 5.69ms | -15.51 kB ± 42.33 kB | 5 |
| Query 3 indexes (1st run) | 757.49ms ± 8.14ms | 397.25 kB ± 0 kB | 5 |
| Query 3 indexes (2nd run) | 276.89ms ± 15.39ms | 295.7 kB ± 0 kB | 5 |
| Load two indexes concurrently | 530.35ms ± 10.39ms | -558.9 kB ± 44.37 kB | 5 |
| Paginate 20000 msgs with pageSize=5 | 9009.28ms ± 3519.35ms | -4.29 MB ± 0 MB | 5 |
| Paginate 20000 msgs with pageSize=500 | 409.66ms ± 33.34ms | -488.18 kB ± 35.92 kB | 5 |
| Query a prefix map (1st run) | 177.39ms ± 1.85ms | 44.16 kB ± 0 kB | 5 |
| Query a prefix map (2nd run) | 9.66ms ± 0.37ms | 7.17 MB ± 0.03 MB | 5 |