@staltz 2021-04-08
@Barbarrosa Thanks for the useful info. It's a bit puzzling why do the benchmarks not look better with this PR.

> db query warmup (10319): 1.013s
> db query author all (8032): 760.335ms
> db query author post msgs (2095): 236.625ms
> db query author 10 msgs: 26.373ms

@arj03 Could you share the code snippets to produce these author benchmarks?