@arj03 2021-04-08
I used server.js from [here](https://github.com/arj03/ssb-sync-test). I can maybe try adding that as a benchmark here.