@arj03 2021-04-08
The reason for running them outside was that I wanted to know how it performs on real live data.