@arj03 2021-04-08
Added a benchmark where this new code is better. Before it was 58ms, now it is 15ms.

I'll try to see if I can figure out why `Query 3 indexes (2nd run)` is slower. That is the one I'm most worried about.