@arj03 2021-04-08
So the problem was that `query 3` basically end up with query that looks like 2000 results that needs to be filter checked and then 20k that don't (the contact type) and before we did the filter check on all of them.